BinomialSets.jl

Documentation for BinomialSets.jl

IPGBs.BinomialSets.BinomialSet — Type

Represents a set of binomials (e.g. a partial or complete Gröbner Basis), including structures for efficient reduction of elements.

In order to allow 4ti2-form bases, we allow elements to be Vector{Int}, in addition to GBElements.

IPGBs.BinomialSets.change_ordering! — Method

Changes the ordering of bs to a monomial order given by the matrix new_order.

IPGBs.BinomialSets.fourti2_form — Method

Returns the binomials of bs as integer vectors in minimization form.

IPGBs.BinomialSets.is_groebner_basis — Method

Returns true iff bs is a Gröbner Basis. This is checked by applying Buchberger's theorem and generating all S-binomials, thus may be slow. Note that this verifies if bs is a Gröbner Basis, but not a truncated Gröbner Basis. To check the latter case, use istruncatedgroebner_basis.

Useful for debugging and automatic tests.

IPGBs.BinomialSets.is_support_reducible — Method

Returns true if (i, j) should be discarded.

This is an implementation of Criteria 1 and 2 from Malkin's thesis.

Criterion 1 is simply the classical Buchberger GCD criterion. If the positive supports are disjoint (= GCD of the leading terms is 1) then the pair may be discarded.

Criterion 2 is specific to homogeneous ideals (or just ideals coming from lattices?) Either way, if the negative supports are not disjoint (= GCD of trailing terms is not 1) then the pair may be discarded. This applies specifically to the bounded components (= variables) of the problem.

I also added new criteria for binary variables. In this case, a pair (i, j) may be discarded if the positive binaries of i are disjoint from the negative binaries of j, and vice-versa. This is because in these cases the pair will generate some binomial with 2 or -2 in some coordinate. These are never necessary in a Gröbner basis.

IPGBs.BinomialSets.is_truncated_groebner_basis — Method

is_truncated_groebner_basis(
bs :: BinomialSet{T, S},
truncate :: Function

) :: Bool where {T <: AbstractVector{Int}, S <: GBOrder}

Returns true if and only if bs is a truncated Gröbner Basis with truncation done using the truncate function.

IPGBs.BinomialSets.minimal_basis! — Method

Updates gb to a minimal Gröbner Basis.

For more info on minimal GBs, see Lemma 3 from Cox, Little, O'Shea Chapter 2.7. In summary, it shows that one can remove from a GB any g such that LT(g) is a multiple of LT(h), for h distinct from g in the GB.

IPGBs.BinomialSets.reduce! — Method

Reduces g by bs efficiently using a Support Tree. Returns a pair of booleans (r, c) where r == true iff g reduced to zero, and c == true iff g was changed in this reduction.

IPGBs.BinomialSets.reduce! — Method

Fully reduce binomial by gb in-place, finding its normal form. Uses tree to speed up the search for reducers. Returns a pair of booleans (r, c) where r == true iff g reduced to zero, and c == true iff g was changed in this reduction.

binomial can also be a monomial.

If reduction_count is given, the number of times each reducer was used in this reduction process is added to reduction_count. This greatly slows down execution, and is only meant for experimental purposes. The parameter should be set to nothing in practical runs.

IPGBs.BinomialSets.reduce! — Method

reduce!(
g :: T,
bs :: Vector{T}

) :: Tuple{Bool, Bool} where {T <: AbstractVector{Int}}

Reduce `g` by `bs` inefficiently by linearly searching for a reducer.
This is useful for very small examples and debugging. For anything else,
consider making a BinomialSet and reducing using a SupportTree.

Return a tuple (r, c) where r == true iff `g` reduced to zero, and c == true
iff `g` was changed in this reduction.

IPGBs.BinomialSets.reduced_basis! — Method

Updates gb to a reduced Gröbner Basis.

IPGBs.BinomialSets.sbinomial — Method

Creates a concrete S-binomial from pair. In practice, this should only be called after we were unable to eliminate pair.