Buchberger.jl

Documentation for Buchberger.jl

IPGBs.Buchberger — Module

Implements a combinatorial Buchberger algorithm for Integer Programming where all data is non-negative. Based on Thomas and Weismantel (1997).

IPGBs.Buchberger.BuchbergerState — Type

The state of Buchberger S-binomial generation.

IPGBs.Buchberger.next_state! — Method

Updates the state to the indices of the next generated S-pair. Returns the indices of the new state, if it corresponds to a new S-pair, or nothing, if all S-pairs were already generated.

IPGBs.GBAlgorithms.late_pair_elimination — Method

Applies the GCD criterion to determine whether or not to eliminate the given S-pair.

IPGBs.GBAlgorithms.next_pair! — Method

Produces the next BinomialPair to be processed by the Buchberger algorithm, if one exists, or nothing otherwise.

Auto-reduces the partial GB periodically (for consistency with 4ti2).

IPGBs.GBAlgorithms.quick_truncation — Method

GBAlgorithms.quick_truncation(
algorithm :: BuchbergerAlgorithm{T},
g :: T

) where {T <: GBElement}

Return true iff g is truncated by the weight criterion, that is, its weight is higher than the maximum weight for fibers of the given instance.