GBAlgorithms.jl

Documentation for GBAlgorithms.jl

IPGBs.GBAlgorithms.GBAlgorithm — Type

A generic Gröbner Basis algorithm. Currently, it can be either a BuchbergerAlgorithm or a SignatureAlgorithm.

IPGBs.GBAlgorithms.general_truncate — Method

general_truncate(algorithm :: GBAlgorithm, binomial :: GBElement) :: Bool

Return true iff binomial is truncated by the weight criterion or by the general truncation method used by algorithm.

IPGBs.GBAlgorithms.increment — Method

increment(algorithm :: GBAlgorithm, stat :: Symbol)

Increment the field stat of stats(algorithm)

IPGBs.GBAlgorithms.is_minimization — Method

is_minimization(algorithm :: GBAlgorithm) :: Bool

Return true iff this algorithm is working with problems in minimization form.

IPGBs.GBAlgorithms.late_pair_elimination — Method

late_pair_elimination(:: GBAlgorithm, :: CriticalPair)

Return true iff the given CriticalPair should be eliminated according to this algorithm's criteria.

IPGBs.GBAlgorithms.prepare_gb_output — Method

prepare_gb_output(algorithm :: GBAlgorithm) :: Vector{Vector{Int}}

Return the user-friendly output Gröbner Basis of algorithm, reduced, sorted and in the original variable ordering.

This output is compatible with 4ti2's output format.

IPGBs.GBAlgorithms.reintroduce_truncated! — Method

reintroduce_truncated!(algorithm :: GBAlgorithm)

Add truncated elements of the given Markov basis back into the GB.

This is used at the end of the computation to keep the GB as a basis of the input ideal, regardless of truncation.

IPGBs.GBAlgorithms.run — Method

run(algorithm :: GBAlgorithm, quiet :: Bool) :: Vector{Vector{Int}}

Return the Gröbner Basis obtained by running algorithm.