GradedBinomials.jl

Documentation for GradedBinomials.jl

IPGBs.GradedBinomials — Module

A structure for efficient computation of truncated GBs of IPs where the data is positive, as described in Section 3 of Thomas and Weismantel (1997). This type of implicit representation of GB elements is also used in Urbaniak, Weismantel and Ziegler (1997).

TODO there are multiple implementations of sparse subtraction of GradedBinomials here. I should probably check whether I need them all

TODO Currently, using GradedBinomials gives different (but apparently correct) GBs than using Binomials. There's probably something implicit about the monomial order they're computed wrt that I'm missing, due to one being maximization and the other being minimization.

Regardless, I'm unconvinced this will ever be more efficient than the usual Binomials. The main reason is that the implicit variables are updated and checked in the reduction process regardless, only this happens in a different way in the SupportTree. A slightly different implementation, where the implicit variables aren't stored anywhere and are computed as needed could be useful if memory became an issue, but that is another thing entirely.

IPGBs.GradedBinomials.GradedBinomial — Type

Represents a binomial as a vector of integers, tracking its leading and trailing terms and degree for easy and efficient implementation of truncated GBs.

Base.:- — Method

Subtracts two GradedBinomials, updating head/tail and cost. This is more efficient than just subtracting directly.

Base.:- — Method

Unary minus for GradedBinomials. It swaps head by tail and inverts cost.

Base.setindex! — Method

Sets g at index i to v. In theory, this shouldn't be used, as GradedBinomials should not be changed.

IPGBs.GBElements.degree_reducible — Method

Checks whether g.degree+ <= h.degree coordinate-wise, where g.degree+ is the set of positive subset of coordinates of g.degree.

If negative == true, checks whether g.degree- <= h.degree-

IPGBs.GBElements.fullform — Method

Given g = v+ - v-, returns the full form of g including slack variables. This is given by: x^(v+) s^((Av)-) r^(v-) - x^(v-) s^((Av)+) r^(v+)

IPGBs.GBElements.minus — Method

TODO do this in a memory-efficient way by reusing result

IPGBs.GBElements.opposite! — Method

In-place unary minus for GradedBinomials

IPGBs.GBElements.reduce! — Method

Computes g -= h in place.

Returns true iff g reduced to zero.

IPGBs.GBElements.supports — Method

Computes bitsets with the positive and negative supports of g.

IPGBs.GradedBinomials.decomposition — Method

Decomposes g = v+ - v-, where both v+ and v- are non-negative arrays.

IPGBs.GradedBinomials.filter — Method

Computes the filter of binomial. In case fullfilter == false, this is just the indices in the support of the leading term of binomial.

Otherwise (fullfilter == true) this is the vector of ordered indices of the support of binomial, so it also includes indices in the trailing term.

IPGBs.GradedBinomials.fourti2_form — Method

Transforms a vector of GradedBinomials into the form of 4ti2 input/output. This includes putting GradedBinomials into fullform, including all of their variables which are implicit in the GradedBinomial structure.

If is_maximization is true, assumes the GradedBinomials are in maximization form. As 4ti2 assumes minimization, signs have to be swapped.

IPGBs.GradedBinomials.lattice_generator_graded — Function

Vector in Z^n with i-th coordinate 1 and remaining coordinates 0.

IPGBs.GradedBinomials.reduce_negative! — Method

Computes g = h - g changing g in-place. This is useful in the case g is supposed to be reduced by h but g.cost < h.cost.