IPInstances.jl

Documentation for IPInstances.jl

IPGBs.IPInstances.IPInstance — Type

Represents an instance of a problem

min C * x

s.t. A * x = b

0 <= x <= u

x in ZZ^n

The instance is stored in normalized form, with permuted variables so that the variables appear in the following order: bounded, non-negative but unbounded, unrestricted.

IPGBs.IPInstances.apply_permutation — Method

apply_permutation(vector_set :: Vector{Vector{Int}}, permutation :: Vector{Int}) :: Vector{Vector{Int}}

Apply permutation to each vector in vector_set.

IPGBs.IPInstances.compute_permutation — Method

compute_permutation(bounded :: Vector{Bool}, nonnegative :: Vector{Bool}) :: Tuple{Vector{Int}, Int, Int}

Return a permutation of variables that puts the variables in the order [ bounded ; unbounded and restricted ; unrestricted ], along with the indices of the last bounded variable and the last unbounded but restricted variable.

This operation should be stable with respect to the initial ordering of variables.

A permutation is represented by a vector perm such that perm[i] = j means that variable i is sent to j by the permutation.

IPGBs.IPInstances.extract_bound — Method

extract_bound(model :: JuMP.Model, c :: JuMP.ConstraintRef, x :: Vector{JuMP.VariableRef})

Extract lower / upper bound value from a VariableRef type constraint c.

IPGBs.IPInstances.extract_constraint — Method

extract_constraint(model :: JuMP.Model, c :: JuMP.ConstraintRef, x :: Vector{JuMP.VariableRef})

Extract numerical coefficients from a JuMP constraint, returning a vector with left-hand side coefficients and the right-hand side value.

Assumes this is a scalar constraint (= a single constraint).

IPGBs.IPInstances.group_relaxation — Method

group_relaxation(instance :: IPInstance) :: IPInstance

Return a new IPInstance corresponding to the relaxation in instance of the non-negativity constraints of the basic variables in the optimal solution of its linear relaxation.

IPGBs.IPInstances.is_bounded — Method

is_bounded(i :: Int, instance :: IPInstance) :: Bool

Return true iff the variable of index i in instance is bounded.

IPGBs.IPInstances.is_nonnegative — Method

is_nonnegative(i :: Int, instance :: IPInstance) :: Bool

Return true iff the variable of index i in instance is nonnegative.

IPGBs.IPInstances.jump_model — Method

initial_jump_model(
instance :: IPInstance,
first_objective :: Int = 1

)

Return a JuMP Model corresponding to the IPInstance.

The JuMP Model keeps all objectives in its objective function matrix.

IPGBs.IPInstances.nonnegative_data_only — Method

nonnegative_data_only(instance :: IPInstance) :: Bool

Return true iff all data in instance.A and instance.b is non-negative and all variables are non-negative.

IPGBs.IPInstances.nonnegative_vars — Method

nonnegative_vars(instance :: IPInstance) :: Vector{Bool}

Return a boolean vector indicating whether each variable of instance is non-negative.

IPGBs.IPInstances.nonnegativity_relaxation — Method

nonnegativity_relaxation(instance :: IPInstance, nonnegative :: Vector{Bool}) :: IPInstance

Return a new IPInstance corresponding to the relaxation of instance consisting of only keeping the non-negativity constraints of variables marked in nonnegative.

IPGBs.IPInstances.normalize_ip — Method

normalize_ip(A :: Matrix{Int}, b :: Vector{Int}, c :: Matrix{T}, u :: Vector{<: Union{Int, Nothing}}, nonnegative :: Vector{Bool}; ...) where {T <: Real}

Transform a problem in the form: max C * x s.t. Ax <= b 0 <= x <= u

to something of the form max C * x s.t. Ax == b x == u

by adding slack variables.

IPGBs.IPInstances.original_variable_order — Method

original_variable_order(vector_set :: Vector{Vector{Int}}, instance :: IPInstance) :: Vector{Vector{Int}}

Invert the variable permutation of instance over vector_set, returning to the original problem's variable order.

This is useful to give users output in the same variable order they input.

IPGBs.IPInstances.random_ipinstance — Method

random_instance(m :: Int, n :: Int) :: IPInstance

Return a random feasible IPInstance with m constraints and n variables.

IPGBs.IPInstances.solve — Method

solve(instance :: IPInstance)

Return the optimal solution to this IPInstance. This solution is computed
using a traditional IP solver.

IPGBs.IPInstances.unboundedness_proof — Method

unboundedness_proof(
instance :: IPInstance,
i :: Int

) :: Vector{Int}

Return a vector u in kernel(instance.A) proving that the variable of index i is unbounded.