Markov.jl

Documentation for Markov.jl

IPGBs.Markov — Module

Implementation of algorithms to compute Markov bases. These are necessary as sets of generators of an ideal for computing Gröbner bases.

Two methods are implemented:

a simplified method that only works when all data defining an IP is negative

the project-and-lift algorithm of 4ti2 / Malkin's thesis

IPGBs.Markov.ProjectAndLiftState — Type

State of the project-and-lift algorithm representing the relevant
information during a specific iteration.

- `original_instance`: the original IPInstance being solved

- `working_instance`: instance including potential upper bounds for efficiency

- `unlifted`: the remaining unlifted variables (original indexing, following
`working_instance`)

- `nonnegative`: the variables that have nonnegative constraints (original indexing)

- `relaxation`: IPInstance representing the problem with the non-negativity of
the unlifted variables relaxed

- `markov`: current partial Markov basis (indexed by the permuted variables)

- `primal_solutions`: feasible solutions for the original problem (and thus all of
its group relaxations). Indexed by the relaxation variables.

- `dual_solution: feasible solution for `relaxation`. Computing a feasible
solution is essentially free in project-and-lift, so we always compute one.
If it is ever feasible for the original problem, it is also optimal for it.
Uses the variable ordering of `relaxation`.

- `optimal_solution`: The optimal solution for the original problem, if known
(original indexing)

- `has_optimal_solution`: Whether the optimal solution is known.

- `stats`: Total statistics for this run of project-and-lift

- `completion`: The GB algorithm to be used. :IPGBs and :4ti2 are supported.

IPGBs.Markov.can_lift — Method

can_lift(markov :: Vector{Vector{Int}}, i :: Int)

Return true if the variable indexed by `i` can be immediately lifted given
the current Markov basis `markov`.

This happens when there is no element in `markov` with a positive coefficient.

IPGBs.Markov.initialize_project_and_lift — Method

initialize_project_and_lift(instance :: IPInstance)

Create the initial state of the project-and-lift algorithm for IP instance, including a Markov basis for its group relaxation.

The group relaxation Markov basis is obtained through the Hermite Normal Form algorithm.

IPGBs.Markov.is_finished — Method

is_finished(pl_state::ProjectAndLiftState)

Check whether the project-and-lift algorithm is ready to terminate at pl_state.

IPGBs.Markov.lift_and_relax — Method

lift_and_relax(
pl_state :: ProjectAndLiftState,
markov :: Vector{Vector{Int}}

) :: ProjectAndLiftState

Lift the variables in `pl_state` according to `markov` generating a new state
with the updated group relaxation, Markov basis and solutions.

IPGBs.Markov.lift_bounded — Method

lift_bounded(
pl_state :: ProjectAndLiftState,
i :: Int

) :: Vector{Vector{Int}}

Lift the variable indexed by `i` in `pl_state` using a Gröbner basis computation.
This generates a Markov basis for the next group relaxation.

IPGBs.Markov.lift_unbounded — Method

lift_unbounded(
pl_state :: ProjectAndLiftState,
i :: Int,
ray :: Vector{Int}

) :: Vector{Vector{Int}}

Lift the variable indexed by `i` in `pl_state` using `ray` that proves that
the variable is unbounded in `pl_state.relaxation`.

IPGBs.Markov.lift_variables! — Method

lift_variables!(
pl_state :: ProjectAndLiftState,
markov :: Vector{Vector{Int}}

)

Mark all unlifted variables that can be lifted in `pl_state` as nonnegative
and lift them.

IPGBs.Markov.markov_basis — Function

Compute a Markov basis of an IP.

IPGBs.Markov.most_negative — Method

most_negative(sigma, solution)

Find the index of the most negative variable in `solution` indexed by `sigma`.

IPGBs.Markov.next — Method

next(
s::ProjectAndLiftState;
completion :: Symbol = :Buchberger,
truncation_type::Symbol = :LP

)::ProjectAndLiftState

Run a single iteration of the project-and-lift algorithm over state, returning the new state after that iteration.

One iteration involves lifting a previously unlifted variable, either through a linear program or a Gröbner Basis computation.

IPGBs.Markov.optimize_relaxation! — Method

optimize_relaxation!(pl_state :: ProjectAndLiftState)

Optimize `pl_state.dual_solution`, which is feasible for `pl_state.relaxation`,
obtaining an optimal solution for the relaxation.

Gröbner bases are used to optimize the solution.

IPGBs.Markov.original_index — Method

original_index(i :: Int, r :: IPInstance)

Index of `r`'s variable `i` in the original instance.

IPGBs.Markov.project_and_lift — Method

project_and_lift(
instance::IPInstance;
completion :: Symbol = :IPGBs,
truncation_type::Symbol = :LP,
optimize :: Bool = false,
feasible :: Bool = false,
quiet :: Bool = true,
solution :: Vector{Int} = zeros(Int, instance.n)

)::Tuple{Vector{Vector{Int}}, Bool, Vector{Int}, Float64}

Compute a Markov basis of instance using the project-and-lift algorithm.

Truncation is done with respect to truncation_type. If truncation_type is None, then a full Markov basis is computed instead.

IPGBs.Markov.relax_and_reorder — Method

relax_and_reorder(
pl_state :: ProjectAndLiftState,
markov :: Vector{Vector{Int}}

) :: Tuple{IPInstance, Vector{Vector{Int}}, Vector{Vector{Int}}, Vector{Int}}

Relax the non-negativity constraints of `pl_state.working_instance` according to
`pl_state.nonnegative`, and reorder the Markov basis and solutions to match the
new variable ordering.

IPGBs.Markov.relaxation_feasible — Method

relaxation_feasible(pl_state :: ProjectAndLiftState)

The feasible solution for the relaxation, in the original variable ordering.

IPGBs.Markov.relaxation_index — Method

relaxation_index(i :: Int, relaxation :: IPInstance)

Index of original variable `i` in `relaxation`.

IPGBs.Markov.simple_markov — Method

simple_markov(instance :: IPInstance) :: Vector{Vector{Int}}

Compute a Markov basis of instance with the simplified algorithm. Assumes all data in instance.A and instance.b is non-negative.

The simplified algorithm uses the fact (proved in, e.g. Thomas and Weismantel (1997)) that the unit vectors on the original variables of the IP form a Markov basis.

IPGBs.Markov.truncate_markov — Method

truncate_markov(
markov::Vector{Vector{Int}},
instance::IPInstance,
truncation_type::Symbol

)::Vector{Vector{Int}}

Compute a subset of markov including only vectors which shouldn't be truncated according to the rule given by truncation_type.