MatrixTools.jl

Documentation for MatrixTools.jl

IPGBs.MatrixTools.hnf_lattice_basis — Method

hnf_lattice_basis(A :: Matrix{Int})

Return a row basis for the lattice ker(A) computed using the Upper
Hermite Normal Form. The entries of this basis tend to be smaller than
those computed directly from kernel(A).

IPGBs.MatrixTools.initial_solution — Method

initial_solution(
A :: Matrix{Int},
b :: Vector{Int}

) :: Vector{Int}

Return a solution to instance.A * x == instance.b, dropping the non-negativity constraints.

IPGBs.MatrixTools.is_normalized_hnf — Method

is_normalized_hnf(H :: Generic.MatSpaceElem{T})

Return true iff H is in normalized HNF form as defined in normalize_hnf!.

IPGBs.MatrixTools.lift_vector — Method

lift_vector(
v :: Vector{Int},
projected_basis :: Generic.MatSpaceElem{Int},
lattice_basis :: Generic.MatSpaceElem{Int}

) :: Vector{Int}

Lift v from an (extended) group relaxation to the full problem given by instance.

IPGBs.MatrixTools.normalize_hnf! — Method

normalize_hnf!(H :: Generic.MatSpaceElem{T})

Change H to an upper row HNF matrix satisfying the following property: all entries above a pivot are non-positive and of smaller magnitude than the pivot

Assumes H is already in HNF as defined by AbstractAlgebra.jl, that is, it is in upper row HNF satisfying:

all entries above a pivot are non-negative and of smaller magnitude than the pivot

IPGBs.MatrixTools.solve — Method

solve(A :: Matrix{Int}, b :: Vector{Int}) :: Vector{Int}

A solution to Ax = b, that is, an element of the fiber of right-hand side
`b` in the lattice ker(A).