MultiObjectiveAlgorithms.jl

Documentation for MultiObjectiveAlgorithms.jl

IPGBs.MultiObjectiveAlgorithms — Module

A module to solve multi-objective integer programming problems using epsilon-constraint methods and test sets / Gröbner Bases.

IPGBs.MultiObjectiveAlgorithms.gbelems_with_positive_slack — Method

gbelems_with_positive_slack(
gb::Vector{Vector{Int}},
slack_indices::Vector{Int}

)::Vector{Vector{Int}}

Return the vectors in gb with no negative value and at least one strictly positive value in the slack_indices variables.

IPGBs.MultiObjectiveAlgorithms.minimal_steps — Method

minimalsteps(
gb::Vector{Vector{Int}},
x::Vector{Int}

)::Vector{Vector{Int}}

Return the set of minimal steps in objective value from x that may yield new points of the Pareto frontier in an epsilon-constraint method.

IPGBs.MultiObjectiveAlgorithms.moip_gb — Method

moip_gb(s :: MOIPGBState) :: BinomialSet{Vector{Int}, MonomialOrder}

Gröbner Basis / test set for the multi-objective problem corresponding to the state s.

The GB can be computed using either 4ti2 or IPGBs, depending on the solver specified in s.

IPGBs.MultiObjectiveAlgorithms.moip_gb_solve — Method

moip_gb_solve(
instance::IPInstance,
initial_solution::Vector{Int};
solver::String="4ti2"

)::Tuple{Vector{Vector{Int}}, Set{Vector{Int}}, Stats}

Pareto front for multi-objective instance given initial_solution, using solver.

If solver is a test-set based solver, applies the project-and-lift algorithm to obtain a solution.

IPGBs.MultiObjectiveAlgorithms.new_step_list — Method

new_step_list(steps :: Vector{Vector{Int}}, slacks :: Vector{Int}) :: Vector{Vector{Int}}

Add the slacks vector to the steps list if it is not dominated by any vector in steps, and remove from steps any vector dominated by slacks.

In the biobjective case this is unnecessary, but dropping useless steps can lead to better performance when there are 3+ objectives.

IPGBs.MultiObjectiveAlgorithms.push_efficient_point! — Method

push_efficient_point!(state :: MOIPGBState, point :: Vector{Int})

Update state's efficient and nondominated sets with point whenever point is a previously unknown efficient solution.

IPGBs.MultiObjectiveAlgorithms.solve_current_moip! — Method

solve_current_moip!(s::MOIPGBState)

Solve the scalarization of a MOIP given by s. If s.solver uses test sets, then a test set is computed, in case one is not already available. Otherwise, s.solver is used directly to solve the scalarization.