MultiObjectiveTools.jl

Documentation for MultiObjectiveTools.jl

IPGBs.MultiObjectiveTools — Module

Some useful functions to deal with multi-objective optimization problems.

IPGBs.MultiObjectiveTools.change_direction! — Method

Changes direction of the given Pareto set by changing signs (i.e., transform a solution from a minimization problem into maximization and vice-versa).

IPGBs.MultiObjectiveTools.contained_by — Method

Checks whether pareto1 is contained by pareto2.

IPGBs.MultiObjectiveTools.is_feasible — Method

is_feasible(x :: Vector{Int}, A :: Matrix{Int}, b :: Vector{Int})

Check whether `x` is feasible for `Ax = b, x \geq 0`.

IPGBs.MultiObjectiveTools.is_nondominated — Method

Checks whether the solution at index i of pareto is nondominated with respect to the rest of pareto.

Returns nothing in case the solution is nondominated, or an index pointing to proof that the solution is inefficient, otherwise.

IPGBs.MultiObjectiveTools.pareto_vector — Method

Creates a vector of pareto points from a set of pareto points

IPGBs.MultiObjectiveTools.print_dominated — Method

For debugging. Prints all dominated points in pareto.