Optimize.jl

Documentation for Optimize.jl

IPGBs.Optimize.optimize_with! — Method

optimize_with(
solution :: Vector{Int},
instance :: IPInstance,
test_set :: Vector{Vector{Int}}

) :: Vector{Int}

Use `test_set` to optimize `solution`, a feasible solution to `instance`,
returning a new feasible solution that cannot be further improved by
`test_set`.

If `test_set` is a Gröbner basis, then the returned solution is optimal.
Otherwise, it is simply a locally optimal solution with respect to `test_set`.