Orders.jl

Documentation for Orders.jl

IPGBs.Orders.GBOrder — Type

This is specialized by MonomialOrder in case of Buchberger's algorithm and by ModuleMonomialOrder in case of Signature-based algorithms.

IPGBs.Orders.MonomialOrder — Type

Implements a monomial order from a given cost matrix, including a grevlex tiebreaker if necessary.

The cost matrix is represented in column-major order to speed up comparisons.

IPGBs.Orders.change_ordering! — Method

change_ordering!(order :: MonomialOrder, new_order :: Matrix{Float64}, A :: Matrix{Int}, b :: Vector{Int})

Change the order stored in MonomialOrder to a new order.

The tiebreaker is rebuilt if necessary.

IPGBs.Orders.invert_maximization — Method

invert_maximization(order :: MonomialOrder, result :: Bool)

Invert result when order is a maximization order.

IPGBs.Orders.is_inverted — Method

is_inverted(
order :: MonomialOrder,
v :: T,
costs :: Vector{Int}

) :: Bool where {T <: AbstractVector{Int}}

Efficiently return whether v's leading and trailing terms are inverted.

IPGBs.Orders.is_inverted_generic — Method

is_inverted_generic(order :: MonomialOrder, v :: T) :: Bool where {T <: AbstractVector{Int}}

Return true iff the trailing and leading terms of v are inverted with respect to order.

IPGBs.Orders.is_inverted_invlex — Method

is_inverted_invlex(order :: MonomialOrder, v :: T) :: Bool where {T <: AbstractVector{Int}}

Return true iff the trailing and leading terms of v are inverted with respect to the invlex order.

v is inverted according to invlex iff its first non-zero entry is positive. Inversion testing is thus done more efficiently in the invlex case.

IPGBs.Orders.lex_order — Method

lex_order(n :: Int) :: MonomialOrder

Return the lex monomial order with x1 > x2 > ... > xn.

IPGBs.Orders.normalize_order — Method

normalize_order(
C::Matrix{Float64},
A::Matrix{Int},
b::Vector{Int},
num_vars::Int,
unbounded::Vector{Bool}

)::Tuple{Matrix{Float64}, Int}

Return a cost matrix in column-major form giving the same monomial order as C using only the first num_vars variables.

Normalization uses the dual data of the LP min C*x s.t. Ax = b, x >= 0.

IPGBs.Orders.order_costs — Method

order_costs(
order :: MonomialOrder,
v :: AbstractVector{Int}

) :: Vector{Int}

The cost of some vector according to the cost matrix of order.

This is the vector weighted by the first column of the cost matrix.

IPGBs.Orders.positive_first_row! — Method

positive_first_row!(C :: Matrix{Float64}, A :: Matrix{Int}, b :: Vector{Int})

Modify C so that its first row is strictly positive using dual data from the LP min C*x s.t. Ax = b, x >= 0.

Assumes the given LP is bounded and feasible.

IPGBs.Orders.projection_order — Method

projection_order(C :: Matrix{Float64}, A :: Matrix{Int}, b :: Vector{Int}, num_vars :: Int)

Return a matrix corresponding to the same monomial order as C but using only the first num_vars variables.

Uses the dual of the LP min C*x s.t. Ax = b, x >= 0 to find such an order.

IPGBs.Orders.tiebreak — Method

tiebreak(C :: Matrix{Float64}) :: Matrix{Float64}

Return a matrix corresponding to the same monomial order as C, tiebreaking it using the revlex order if C does not already specify a full monomial order itself.