SignaturePolynomials.jl

Documentation for SignaturePolynomials.jl

IPGBs.SignaturePolynomials — Module

Implementation of the basic signature-polynomial pairs structures.

TODO this whole module would probably be cleaner if it was separated in multiple modules. One defining signatures, another defining module monomial orders, another defining SigPolys and so on.

IPGBs.SignaturePolynomials.ModuleMonomialOrder — Type

An enum containing all types of module monomial orders available for this signature algorithm implementation.

IPGBs.SignaturePolynomials.ModuleMonomialOrdering — Type

Represents a module monomial ordering. This has to store a reference to the current basis in order to implement an order like lt-pot (Schreyer's order).

This is mutable because, for convenience, it may sometimes be necessary to change monomial_order.

IPGBs.SignaturePolynomials.SigPoly — Type

Represents a polynomial with its signature. Used in signature-based algorithms. Stores its sig-lead ratio to compare signatures more efficiently when building S-pairs, Koszul syzygies, reductions and elimination criteria.

IPGBs.SignaturePolynomials.Signature — Type

A signature, or module monomial, is represented by a monomial with an index, where the monomial is the coefficient of the module basis vector of the given index.

TODO I'm thinking that monomial could easily be a sparse vector in cases with many variables. I expect many entries to be 0 in this case, but it is worth it to experiment with that.

IPGBs.SignaturePolynomials.SignaturePair — Type

An S-pair represented sparsely, before building it from binomials explicitly. Includes a signature to allow the signature-based algorithm to proceed by increasing signature of S-pairs.

Base.Order.lt — Method

Compares two SigPolys as polynomials wrt to the monomial order of o. Ignores signatures and sig-leads.

Base.Order.lt — Method

Compares two signatures with respect to the given ModuleMonomialOrdering. Returns true iff s1 is strictly smaller than s2 with respect to o.

TODO I wonder if I shouldn't also store the weighted versions of signature.monomial This could make all orders more efficient, the obvious tradeoff being memory.

IPGBs.GBElements.build — Method

Builds u - v with signature given by pair.

IPGBs.GBElements.degree_reducible — Method

Checks a reduction condition for implicit variables in the case T == GradedBinomial. Currently unused, because GradedBinomials are not supported with signatures.

IPGBs.GBElements.reduce! — Method

Computes g -= h in-place, assumes this is a regular reduction, so that sig(g - h) = sig(g).

Returns true iff g reduced to zero.

IPGBs.GBElements.signature_reducible — Method

Checks whether g is signature reducible by a reducer with reducer_sig. This means the reducer (after multiplying by a lt quotient) has lower signature than g wrt the module monomial ordering.

IPGBs.GBElements.singular_top_reducible — Method

Checks whether g is singular top-reducible by some reducer with signature reducer_sig. Assumes this reducer divides g.

TODO singulartopreducible and signaturereducible both compare the same signatures. I could call the same function signaturecompare in both! In fact, given the situation I use this, I should just call signature_compare there and be done with it!

IPGBs.SignaturePolynomials.divides — Method

Returns true iff s1 divides s2.

IPGBs.SignaturePolynomials.image_leading_term — Method

Returns the leading term image of s by the function mapping the module basis element of index i to the i-th element of generators.

IPGBs.SignaturePolynomials.koszul — Method

Returns the Koszul signature of the pair of gb indexed by (i, j), if it is regular. Otherwise, returns nothing.

IPGBs.SignaturePolynomials.ltpot_compare — Method

Compares s1 and s2 in Schreyer's order (aka ltpot). This means first comparing the leading terms of the images of the signatures in the polynomial ring and then breaking ties by the pot order.

Returns :lt, :eq or :gt if s1 is smaller, equal or greater than s2 respectively.

TODO try to implement this more efficiently. The way I'm doing this here is very slow, due to all the calls to imageleadingterm...

IPGBs.SignaturePolynomials.ltpot_lt — Method

Returns true iff s1 < s2 in the ltpot (Schreyer) order.

IPGBs.SignaturePolynomials.pot_compare — Method

Compares s1 and s2 in the position-over-term (pot) order, that is, compare signatures first by module index and break ties by the given monomial order on the coefficients.

IPGBs.SignaturePolynomials.pot_lt — Method

Returns true iff s1 < s2 in the top order.

IPGBs.SignaturePolynomials.regular_spair — Method

Creates an SPair S(i, j) if it is regular, otherwise returns nothing.

IPGBs.SignaturePolynomials.siglead — Method

Computes the SigLead ratio of binomial with signature. This is defined as a signature with the same index as signature, but with monomial defined by signature.monomial - leading_term(binomial)

IPGBs.SignaturePolynomials.signature_compare — Method

Returns :lt, :eq or :gt when sig1 is respectively smaller, equal or greater than sig2 with respect to module_order.

IPGBs.SignaturePolynomials.signature_lt — Method

Returns sig1 < sig2 according to the given module_order

IPGBs.SignaturePolynomials.spair_signature — Method

Computes the signature of the i-branch of the S-pair (i, j). Just call with i and j inverted to compute the same thing for j.

IPGBs.SignaturePolynomials.top_compare — Method

Compares s1 and s2 in the term-over-position (top) order, that is, compare signature first by monomial order on the coefficients breaking ties by the module index.

Returns :lt, :eq or :gt if s1 is smaller, equal or greater than s2 respectively.

IPGBs.SignaturePolynomials.top_lt — Method

Returns true iff s1 < s2 in the top order.