SingleObjective.jl

Documentation for SingleObjective.jl

IPGBs.SingleObjective — Module

Tools for solving single objective integer programs given by their model min c' * x s.t. A * x = b x >= 0

IPGBs.SingleObjective.lexmin — Method

lexmin(
A :: Matrix{Int},
b :: Vector{Int},
C :: Matrix{Int},
i :: Int;
solver :: String = "Cbc"

) :: Tuple{Vector{Int}, Int}

Solve the lexicographical minimization of the multiobjective problem

min. C * x s.t. A * x == b x >= 0, x integer

considering the i-th objective as the first objective, and optimizing the remaining objectives as tiebreakers in order 1, 2, ..., i - 1, i + 1, ... p.

IPGBs.SingleObjective.linear_relaxation — Method

Solves the linear relaxation of the integer program min c' * x s.t. A * x = b x >= 0 using JuMP.

IPGBs.SingleObjective.makemodel — Method

Creates a JuMP model (for the given solver) from A, b and c.

IPGBs.SingleObjective.usesolver — Method

Uses JuMP / Cbc to solve an integer programming problem given in the form min c' * x s.t. A * x = b x >= 0