SupportTrees.jl

Documentation for SupportTrees.jl

IPGBs.SupportTrees — Module

Reimplementation of 4ti2's FilterReduction tree structure used to store binomial sets for efficient reductions. Called SupportTrees in reference to Roune and Stillman (2012).

IPGBs.SupportTrees.SupportNode — Type

Each binomial in the GB is inserted in some node of the tree.

A filter is a list of indices of the binomials such that their coordinates are positive - that is, the list of indices of variables appearing in its leading term.

A binomial is inserted in a node of the tree with the same filter as itself. All binomials with the same filter are inserted in the same node. If the tree corresponds to a GB (or a partial GB in some way) each node should contain at most one binomial in its binomial_list.

Each child node is labeled by its (next) filter index.

IPGBs.SupportTrees.SupportTree — Type

The structure from 4ti2's FilterReduction for efficient reduction of binomials in toric ideals.

At each level i of the tree, there exists a node labeled j if there is some binomial in the tree whose filter's i-th element is j. Each binomial is stored at level n, with n the size of its filter.

If fullfilter is true, then both positive and negative entries of the binomials are considered to be in the filter for reduction. This is useful to implement the specialized truncated GB algorithm described in Thomas and Weismantel, Section 3.

IPGBs.SupportTrees.SupportTree — Method

Puts references to the elements of gb in a reduction tree, so that one may easily find whether a given binomial can be reduced by gb.

IPGBs.SupportTrees.add_binomial! — Method

Adds a (reference to a) binomial to a SupportTree.

The idea is adding the binomial to a node corresponding precisely to its filter. At each iteration, take the i-th element of the filter of binomial. If there is already a node in the tree at level i labeled by the i-th index in the filter of binomial, move to the subtree defined by that node. Otherwise, create such a node.

IPGBs.SupportTrees.enumerate_reducers! — Method

Walks in the tree recursively pushing any reducers found to reducers.

IPGBs.SupportTrees.enumerate_reducers — Method

Returns a vector containing all reducers of g in gb, using tree to find them efficiently.

IPGBs.SupportTrees.find_reducer — Method

Checks whether g is reducible by some element of gb and, if so, returns a reducer. Otherwise, returns nothing.

This is done by looking up a node in this subtree with a subfilter of g's filter and then checking directly whether some binomial in the node reduces g.

The lookup operation traverses the tree looking for nodes whose labels are contained in the filter of g.

IPGBs.SupportTrees.find_reducer — Method

Checks whether g is reducible by some element of gb and, if so, returns a reducer. Otherwise, returns nothing.

IPGBs.SupportTrees.remove_binomial! — Method

Removes binomial from tree by doing a lookup operation and then just removes it from the corresponding binomial_list.

If binomial is not in tree, nothing will happen.