TriangleHeaps.jl

Documentation for TriangleHeaps.jl

IPGBs.TriangleHeaps — Module

Implements the S-pair triangles described in Roune and Stillman (2012), Section 4.3.

IPGBs.TriangleHeaps.TriangleHeap — Type

Stores a heap of SignaturePairs (i, j) containing only the S-pair of smallest signature for each i. Then, for each i, a row of triangle contains all the values of j such that (i, j) is still in the queue, without storing the signatures of these S-pairs. This reduces memory consumption.

For further improvements in memory consumption, we use the type parameter I to allow for smaller unsigned integer types. Using Uint16 allows computing with GBs of up to 65535 elements.

Base.isempty — Method

Returns true iff heap is empty. Just delegates to its internal heap, which is only empty when heap is.

Base.pop! — Method

Returns the SignaturePair of lowest signature in heap. Assumes that heap is non-empty.

IPGBs.TriangleHeaps.push_batch! — Method

Puts a batch of SignaturePairs (all coming from pairs (i, j) with fixed j) in heap, adding a new row to its triangle. Has to sort the S-pairs by signature, so its complexity is O(S * n log n) where S is the complexity of signature comparison and n = length(batch).