BinomialSets.jl

Documentation for BinomialSets.jl

IPGBs.BinomialSets.BinomialSet — Type

Represents a set of binomials (e.g. a partial or complete Gröbner Basis), including structures for efficient reduction of elements.

In order to allow 4ti2-form bases, we allow elements to be Vector{Int}, in addition to GBElements.

IPGBs.BinomialSets.auto_reduce_once! — Method

Reduces each element of the GB by the previous elements.

IPGBs.BinomialSets.change_ordering! — Method

Changes the ordering of bs to a monomial order given by the matrix new_order.

IPGBs.BinomialSets.fourti2_form — Method

Returns the binomials of bs as integer vectors in minimization form.

IPGBs.BinomialSets.is_groebner_basis — Method

Returns true iff bs is a Gröbner Basis. This is checked by applying Buchberger's theorem and generating all S-binomials, thus may be slow. Note that this verifies if bs is a Gröbner Basis, but not a truncated Gröbner Basis. To check the latter case, use istruncatedgroebner_basis.

Useful for debugging and automatic tests.

IPGBs.BinomialSets.is_negative_disjoint — Method

Returns true iff the i-th and j-th vectors of bs have disjoint negative bounded components.

This is essentially 4ti2's cancellation criterion.

IPGBs.BinomialSets.is_positive_disjoint — Method

Returns true iff the i-th and j-th vectors of bs have disjoint positive non-negative components.

This is essentially the GCD criterion.

IPGBs.BinomialSets.is_support_reducible — Method

Returns true if (i, j) should be discarded.

This is an implementation of Criteria 1 and 2 from Malkin's thesis.

Criterion 1 is simply the classical Buchberger GCD criterion. If the positive supports are disjoint (= GCD of the leading terms is 1) then the pair may be discarded.

Criterion 2 is specific to homogeneous ideals (or just ideals coming from lattices?) Either way, if the negative supports are not disjoint (= GCD of trailing terms is not 1) then the pair may be discarded. This applies specifically to the bounded components (= variables) of the problem.

IPGBs.BinomialSets.is_truncated_groebner_basis — Method

Checks whether bs is a truncated Gröbner Basis where truncation is done with respect to

Ax <= b x <= u

IPGBs.BinomialSets.minimal_basis! — Method

Updates gb to a minimal Gröbner Basis.

For more info on minimal GBs, see Lemma 3 from Cox, Little, O'Shea Chapter 2.7. In summary, it shows that one can remove from a GB any g such that LT(g) is a multiple of LT(h), for h distinct from g in the GB.

IPGBs.BinomialSets.reduce! — Method

Reduces g by bs efficiently using a Support Tree. Returns a pair of booleans (r, c) where r == true iff g reduced to zero, and c == true iff g was changed in this reduction.

IPGBs.BinomialSets.reduce! — Method

Fully reduce binomial by gb in-place, finding its normal form. Uses tree to speed up the search for reducers. Returns a pair of booleans (r, c) where r == true iff g reduced to zero, and c == true iff g was changed in this reduction.

binomial can also be a monomial.

If reduction_count is given, the number of times each reducer was used in this reduction process is added to reduction_count. This greatly slows down execution, and is only meant for experimental purposes. The parameter should be set to nothing in practical runs.

IPGBs.BinomialSets.reduced_basis! — Method

Updates gb to a reduced Gröbner Basis.

IPGBs.BinomialSets.sbinomial — Method

Creates a concrete S-binomial from pair. In practice, this should only be called after we were unable to eliminate pair.