FGLM.jl

Documentation for FGLM.jl

IPGBs.FGLM.direct_normal_form — Method

direct_normal_form(m :: WeightedMonomial, gb :: BinomialSet{T, S}, target_order :: S) where {T, S <: GBOrder}

Compute the normal form of m with respect to gb by applying the usual reduction process directly.

IPGBs.FGLM.fast_normal_form — Method

fast_normal_form(m :: WeightedMonomial, gb :: BinomialSet{T, S}, target_order :: S) where {T, S <: GBOrder}

Compute the normal form of m with respect to gb by computing NormalForm(xi * NormalForm(m')) where m = xi * m'.

This is the method suggested in the original FGLM paper Faugère et al (1994).

IPGBs.FGLM.fglm — Method

fglm(gb1 :: BinomialSet{T, S}, target_order :: S) where {T, S <: GBOrder}

Convert a Gröbner basis gb1 to target_order using the FGLM algorithm.

FGLM only works when the ideal generated by gb1 is zero-dimensional. This is assumed in the implementation.

IPGBs.FGLM.find_linear_dependency — Method

find_linear_dependency(nf :: Vector{Int}, std_basis :: Dict{Vector{Int}, Vector{Int}}

Return an element of std_basis with normal form nf, if such an element exists, otherwise return [].

TODO It is likely possible to do this more efficient using some special kind of tree. I'll leave this for later, in case it's needed.

IPGBs.FGLM.is_below_staircase — Method

is_below_staircase(m :: Vector{Int}, gb :: Vector{T}) where {T}

Checks whether m is below the staircase given by gb. Naïve implementation. It is possible to do this more efficiently by counting the number of times FGLM generated m.

IPGBs.FGLM.is_below_staircase_fast — Method

is_below_staircase_fast(m :: WeightedMonomial)

Check whether m is below a GB staircase indirectly by counting how many times it was inserted in the priority queue.

This idea comes straight from the original FGLM paper Faugère et al (1994).

IPGBs.FGLM.update_monomial_heap! — Method

update_monomial_heap!(h :: MonomialHeap{T}, m :: Vector{Int}) where {T <: GBOrder}

Add all monomials of the form xi * m for all variables xi to the heap h and count repetitions.