FastBitSets.jl

Documentation for FastBitSets.jl

IPGBs.FastBitSets — Module

Specialized module for faster bitsets in the context of GBs / IP.

IPGBs.FastBitSets.BitTriangle — Type

A BitTriangle stores a bit of data for each pair (i, j) with i != j. Data can be set and accessed for both (i, j) and (j, i), even though it is only stored once per pair.

Base.getindex — Method

Truth value associated by this bitset to the i-th variable.

IPGBs.FastBitSets.add_row! — Method

Adds a new row (initialized to false) to this BitTriangle. If this is the n-th row, it will have (n-1) elements.

IPGBs.FastBitSets.check_indices — Method

Checks whether the indices i, j can be a pair in a bit triangle. In practice, this means they have to be different.

IPGBs.FastBitSets.triangle_indices — Method

Swaps indices so that they are access a valid pair in a bit triangle.