FastComparator.jl

Documentation for FastComparator.jl

IPGBs.FastComparator — Module

Implements fast comparisons of complex objects by building a structure of integers representing them.

TODO currently repeated elements give different magnitudes. This is not necessarily a problem yet, but I should fix that if it ever becomes a problem

IPGBs.FastComparator.Comparator — Type

A structure allowing for O(1) comparison of elements of data. Updating the structure with new elements of data can take up to O(n). This is done by associating an integer to each element of data, such that data[i] < data[j] iff magnitudes[i] < magnitudes[j].

TODO Maybe it is possible to optimize insertion to O(log n). Think about it later.

Base.Order.lt — Method

Returns true iff the element of index index1 is smaller than that of index index2 in comp.data. Uses Comparator logic for comparison in O(1).

IPGBs.FastComparator.build! — Method

Returns a vector of sorted permutation indices and integer magnitudes of elements of data in the range given by [-range, range].

IPGBs.FastComparator.check_increasing_property — Method

Checks whether comp satisfies the increasing property, i.e. its elements in the order given by comp.sorted_permutation are in increasing order

This is useful for testing and debugging.

IPGBs.FastComparator.compare — Method

Returns :lt, :eq, :gt if the element indexed by index1 is smaller, equal or greater than the one indexed by index2 respectively. Uses Comparator logic for comparison in O(1).

IPGBs.FastComparator.find_position — Method

Do a binary search for element on the data of comp to find the position where it should be inserted.

IPGBs.FastComparator.increase_range! — Method

Increases the comparator range for magnitudes by RANGEUPDATEFACTOR when possible. If this is not possible, simply increases it to the maximum range possible (full integers).

IPGBs.FastComparator.ith_element — Method

Returns the i-th element of the (sorted) data in comp.

IPGBs.FastComparator.magnitude — Method

Computes the magnitude of the i-th smallest element out of n elements with the given range. The elements are equally spaced in the range.

IPGBs.FastComparator.occupation_ratio — Method

Ratio of occupied slots in [-comp.range, comp.range]

IPGBs.FastComparator.rebuild! — Method

Increases range and rebuilds comp, equally spacing the magnitudes of the elements in the new increased range.

IPGBs.FastComparator.update! — Method

Updates comp with the magnitudes and sorted positions of any new elements added to comp.data. May rebuild the whole structure whenever necessary.

TODO is it possible to lower the complexity of this function from O(n) to O(log n) by avoiding the insertions in a vector? Maybe using some other kind of structure to store comp.sorted_permutation

IPGBs.FastComparator.update_magnitude! — Method

Inserts the magnitude of the element of rank rank in the magnitude list of comp. May rebuild the whole structure if there are no available magnitudes.