FourTi2.jl

Documentation for FourTi2.jl

IPGBs.FourTi2 — Module

Basic Julia-4ti2 interface using system calls. Currently includes the minimize, groebner, normalform, markov and graver 4ti2 commands.

IPGBs.FourTi2._4ti2_clear — Method

Internal use.

Removes the temporary files used to communicate between Julia and 4ti2.

IPGBs.FourTi2._4ti2_read — Method

Internal use.

Reads a Julia array from a given file written by 4ti2.

IPGBs.FourTi2._4ti2_write — Method

Internal use.

Writes a Julia array to a file in a format that can be read by 4ti2.

IPGBs.FourTi2.graver — Function

Calls 4ti2's graver command.

TODO Implement other options (lower and upper bound, arbitrary signs, support to lattice bases instead of matrices...)

IPGBs.FourTi2.groebner — Method

Interface with the 4ti2 groebner command. Computes the Gröbner Basis of the toric ideal given by A using c as the ordering.

If lattice is true, A is a lattice basis instead.

Returns the Gröbner Basis as a matrix where each row corresponds to a pure binomial in it.

IPGBs.FourTi2.groebnernf — Method

Calls 4ti2's groebner command followed by normalform. This is equivalent to minimize, except it generates a Gröbner basis in as output and doesn't apply the optimizations of the latter command.

Returns the optimal solution to the given IP along with its objective value.

IPGBs.FourTi2.markov — Method

Returns the minimal Markov basis of A with respect to c as a matrix, where the rows are the elements of the basis.

IPGBs.FourTi2.minimize — Method

Interface with the 4ti2 minimize command. Solves the Integer Programming problem:

min c' * x s.t. A * x = b x >= 0

where b = A * xinit using Gröbner Basis methods / 4ti2. This is essentially equivalent to calling groebner followed by normalform, but a few optimizations are applied (group relaxations + upper bounds)

I found that –precision=arb is often necessary here to avoid errors.

IPGBs.FourTi2.normalform — Method

Interface with the 4ti2 normalform command. Computes the normal form of xinit with respect to the Gröbner Basis of the toric ideal given by A using c as an ordering. This can be called multiple times for various xinit without recomputing the Gröbner Basis.

Returns the normal form and its value using c as objective function.

IPGBs.FourTi2.write_4ti2_input — Method

Writes an integer programming problem in 4ti2-readable format.

IPGBs.FourTi2.write_4ti2_sign — Function

Writes sign file for 4ti2.