GBElements.jl

Documentation for GBElements.jl

IPGBs.GBElements — Module

This module defines all the binomial data structures used in my implementations of Buchberger's algorithm and Signature-based algorithms.

TODO make GBElements a consistent interface

IPGBs.GBElements.GBElement — Type

Abstract type used for GB computations. It is meant to generalize both binomials and binomials with signature, in order to simplify the implementation of reduction algorithms.

IPGBs.GBElements.degree_reducible — Method

This is only relevant when we consider the implicit representation of binomials. For this explicit representation, we can always return true.

IPGBs.GBElements.filter — Method

Gets the filter of a binomial, that is, the list of indices of variables appearing in its leading term.

If fullfilter = true, include the indices of variables appearing in its trailing term.

IPGBs.GBElements.head — Method

The indices of the positive support of g. Or indices of the support of leading_term(g)

TODO could probably turn this into an iterator instead, would be more efficient

IPGBs.GBElements.ipgbs_form — Method

Turns a 4ti2 GB into my GB format as a vector of integer vectors.

IPGBs.GBElements.is_negative_disjoint — Method

Checks whether the trailing terms of g and h are disjoint. In case negative = true, checks whether the leading term of g is disjoint with the trailing term of h.

IPGBs.GBElements.is_zero — Method

Returns true iff this binomial is zero, that is, all of its coordinates are 0.

IPGBs.GBElements.le_coordinatewise — Method

Returns true iff v1 <= v2 coordinate-wise.

IPGBs.GBElements.le_upperbound — Method

Checks whether v is bounded coordinate by coordinate by u.

IPGBs.GBElements.leading_term — Method

Computes the leading term of this GBElement as a vector.

IPGBs.GBElements.model_truncation — Method

Checks feasibility of:

Ax = b - Av 0 <= x <= u

with x either a vector of real variables or integer variables, whatever is defined in model. In order to do this efficiently, updates the RHS of the previously built model by modifying constraints.

Returns true iff v is feasible for the above model, which means it should NOT be truncated.

TODO LP truncation is slightly slower than I hoped. IP truncation is very slow. Can I implement the model more efficiently somehow?

IPGBs.GBElements.reduce! — Method

Reduce binomial by reducer as many times as possible. Assumes reducer reduces binomial at least once.

Returns true iff binomial reduced to 0.

IPGBs.GBElements.reduces — Method

Checks whether reducer divides g, using the filter of reducer for efficiency. When fullfilter = true, checks if g.head >= reducer.head and g.tail >= reducer.tail coordinate-wise, while also checking a degree criterion.

IPGBs.GBElements.reduction_factor — Method

Finds the maximum k such that k * reducer <= binomial coordinate-wise.

IPGBs.GBElements.simple_truncation — Method

Returns true iff v should be considered for reduction in a truncated GB algorithm.

IPGBs.GBElements.supports — Method

Computes bitsets with positive and negative supports of g.

IPGBs.GBElements.supports — Method

Computes bitsets with positive and negative supports of every element in B.

IPGBs.GBElements.to_gbelement — Method

Turns a vector v into a GBElement of type S. Currently, Binomials are the only subtype supported.

IPGBs.GBElements.truncate — Method

Returns true iff binomial should be truncated according to the given truncation_type.