GBTools.jl

Documentation for GBTools.jl

IPGBs.GBTools.grevlex_matrix — Method

Returns a matrix representing the grevlex order for n variables with xn > x{n-1} > ... > x_1

IPGBs.GBTools.lex_matrix — Method

Returns a matrix representing the lex order for n variables with x1 > x2 > ... > x_n.

IPGBs.GBTools.revlex_matrix — Method

Returns a matrix representing the lex order for n variables with x1 < x2 < ... < x_n.

This is the tiebreaking order used by 4ti2.