Markov.jl

Documentation for Markov.jl

IPGBs.Markov — Module

Implementation of algorithms to compute Markov bases. These are necessary as sets of generators of an ideal for computing Gröbner bases.

Two methods are implemented:

a simplified method that only works when all data defining an IP is negative

the project-and-lift algorithm of 4ti2 / Malkin's thesis

TODO implement a Project-and-Lift state structure. This will facilitate the implementation of group relaxations and early termination, as well as iteratively computing parts of the GB and continuing later

IPGBs.Markov.is_normalized_hnf — Method

is_normalized_hnf(H :: Generic.MatSpaceElem{T})

Return true iff H is in normalized HNF form as defined in normalize_hnf!.

IPGBs.Markov.lex_groebner_basis — Method

lex_groebner_basis(instance :: IPInstance) :: BinomialSet

Compute a lex Gröbner basis of instance using the HNF algorithm.

It is assumed the toric ideal of instance.A is zero-dimensional, i.e. that the lattice optimization problem given by instance is of rank n.

IPGBs.Markov.markov_basis — Function

Compute a Markov basis of an IP.

IPGBs.Markov.normalize_hnf! — Method

normalize_hnf!(H :: Generic.MatSpaceElem{T})

Change H to an upper row HNF matrix satisfying the following property: all entries above a pivot are non-positive and of smaller magnitude than the pivot

Assumes H is already in HNF as defined by AbstractAlgebra.jl, that is, it is in upper row HNF satisfying:

all entries above a pivot are non-negative and of smaller magnitude than the pivot

IPGBs.Markov.project_and_lift — Method

project_and_lift(instance :: IPInstance, truncation_type :: Symbol) :: Vector{Vector{Int}}

Compute a Markov basis of instance using the project-and-lift algorithm.

Truncation is done with respect to truncation_type. If truncation_type is None, then a full Markov basis is computed instead.

TODO I probably could make projections more efficient. But at least this way they are very simple to implement...

IPGBs.Markov.simple_markov — Method

simple_markov(instance :: IPInstance) :: Vector{Vector{Int}}

Compute a Markov basis of instance with the simplified algorithm. Assumes all data in instance.A and instance.b is non-negative.

The simplified algorithm uses the fact (proved in, e.g. Thomas and Weismantel (1997)) that the unit vectors on the original variables of the IP form a Markov basis.

IPGBs.Markov.truncate_markov — Method

truncate_markov(markov :: Vector{Vector{Int}}, instance :: IPInstance, truncation_type :: Symbol) :: Vector{Vector{Int}}

Compute a subset of markov including only vectors which shouldn't be truncated according to the rule given by truncation_type.