SolverTools.jl

Documentation for SolverTools.jl

IPGBs.SolverTools — Module

Includes various LP and IP functions using external solvers and JuMP.

IPGBs.SolverTools.bounded_objective — Method

bounded_objective(A :: Matrix{Int}, i :: Int, sigma :: Vector{Int})

Compute the objective function for the bounded case of the project-and-lift algorithm.

By Farkas Lemma, either A^σ x^σ + Aσ xσ = 0 has a solution with b^T x < 0 or y^T A^σ <= (b^σ)^T y^T Aσ == bσ^T has a solution. In this case of P&L, the latter holds and if y is a solution it follows that c = b - A^T y satisfies: cσ = 0 and c^T u = -ui for all u in ker(A).

IPGBs.SolverTools.feasibility_model — Method

feasibility_model(A :: Matrix{Int}, b :: Vector{Int}, u :: Vector{<: Union{Int, Nothing}}, nonnegative :: Vector{Bool}, var_type :: DataType)

Return a feasibility checking model along with variable and constraint vectors for Ax = b, 0 <= x <= u, where x is either an integer variable vector, if var_type == Int or a real variable vector otherwise.

IPGBs.SolverTools.is_bounded — Method

is_bounded(i :: Int, model :: JuMP.Model, x :: Vector{JuMP.VariableRef})

Return true iff maximizing x_i in model is bounded (has an optimal solution).

It is assumed that the given model is (primal) feasible.

IPGBs.SolverTools.is_bounded — Method

is_bounded(model :: JuMP.Model)

Return true iff the IP/LP given by model is feasible.

IPGBs.SolverTools.is_degenerate — Method

is_degenerate(model :: JuMP.Model, vars :: Vector{VariableRef}, constraints :: Vector{ConstraintRef})

Return true iff the optimal basis of model is degenerate, i.e. if any of its basic variables has value 0.

IPGBs.SolverTools.is_feasible — Method

is_feasible(model :: JuMP.Model)

Return true iff the IP/LP given by model is feasible.

IPGBs.SolverTools.jump_model — Method

jump_model(A :: Matrix{Int}, b :: Vector{Int}, C :: Array{Float64}, u :: Vector{Union{Int, Nothing}}, nonnegative :: Vector{Bool}, var_type :: DataType)

Return a JuMP model (alongside references to its variables and constraints) for the given IP or LP problem.

The optimization problem considered is min c[1, :] * x s.t. Ax = b, x_i >= 0 for all i s.t. nonnegative[i] == true

The variables are integer if var_type == Int or real otherwise.

TODO Should I do anything special for the binary case?

IPGBs.SolverTools.optimal_basis! — Method

optimal_basis!(model :: JuMP.Model) :: Vector{Bool}

Return a boolean vector with value true at index i iff the i-th variable is basic at the optimal solution of model.

This function calls optimize! on model in order to be self-contained.

IPGBs.SolverTools.optimal_row_span — Method

optimal_row_span(A :: Matrix{Int}, b :: Vector{Int}, c :: Array{T}) where {T <: Real}

Compute a vector in the row span of A given by yA where y is the optimal solution to the dual of the LP given by A, b, c.

Assumes Ax = b is feasible, and max x s.t. Ax = b, x >= 0 is bounded. Given these conditions, the dual variables of the constraints of the above LP give a positive row span vector.

IPGBs.SolverTools.positive_row_span — Method

positive_row_span(A :: Matrix{Int}, b :: Vector{Int})

Compute a strictly positive vector in the row span of A using linear programming.

IPGBs.SolverTools.relaxation_model — Method

relaxation_model(A :: Matrix{Int}, b :: Vector{Int}, C :: Array{Float64}, u :: Vector{Union{Int, Nothing}}, nonnegative :: Vector{Bool})

Return a linear relaxation model of the given IP.

The optimization problem considered is of the form min C[1, :]' * x s.t. Ax = b, x >= 0.

IPGBs.SolverTools.set_jump_objective! — Method

set_jump_objective!(model :: JuMP.Model, direction :: Symbol, c :: Vector{T}, x :: Vector{JuMP.VariableRef}) where {T <: Real}

Updates model changing its objective function to c and sense to direction, which can be either :Max or :Min.

IPGBs.SolverTools.unboundedness_ip_model — Method

unboundedness_ip_model(A :: Matrix{Int}, nonnegative :: Vector{Bool}, i :: Int)

Return an IP that checks for the existence of an integer vector u such that Au = 0, u[nonnegative] >= 0, u[i] > 0.

TODO this could also be done with LP as follows: "Assume all data is rational. Then, the polyhedron is rational, so the optimum must be rational. Multiply by a large enough integer..." Implement it this way later!

IPGBs.SolverTools.update_feasibility_model_rhs — Method

update_feasibility_model_rhs(constraints :: Vector{ConstraintRef}, A :: Matrix{Int}, b :: Vector{Int}, v :: T) where {T <: AbstractVector{Int}}

Change the RHS of constraints to b - A * v.