StandardDecomposition.jl

Documentation for StandardDecomposition.jl

IPGBs.StandardDecomposition.AdmissiblePair — Type

Represents an admissible pair, i.e., (x^m, σ) where x^m is a standard monomial and σ is a set of indices disjoint from the support of m.

Additionally, for this pair to be standard, it has to be minimal w.r.t. the following order: (x^m, σ) <= (x^m', σ') iff x^m divides x^m' and supp(x^m' / x^m) ∪ σ' ⊆ σ.

The minimal elements w.r.t. this order are the standard pairs and they cover other admissible pairs.

IPGBs.StandardDecomposition.choose_generator — Method

Chooses some generator of the monomial ideal M. For now, it simply picks the first one.

TODO If I ever care about the performance of the standard decomposition, I can try to study whether it matters which generator is picked or not.

IPGBs.StandardDecomposition.choose_variable — Method

Chooses some variable in the support of monomial. For now, it simply picks the first one.

TODO The same comments from choose_generator apply here.

IPGBs.StandardDecomposition.ideal_quotient — Method

Quotient of (M : (gen / x^i)). The generators of the quotient are the generators m of M divided by gcd(m, gen / x^i)

IPGBs.StandardDecomposition.is_reducible — Method

True iff some generator in gens reduces the i-th element of gens.

IPGBs.StandardDecomposition.is_standard — Method

pair is_standard when it is not dominated by any already known standard pairs.

IPGBs.StandardDecomposition.less_than — Method

Implements the following order: (x^m, σ) <= (x^m', σ') iff x^m divides x^m' and supp(x^m' / x^m) ∪ σ' ⊆ σ.

IPGBs.StandardDecomposition.minimal_generators — Method

Reduces a set of generators gens to a set of minimal generators of the same monomial ideal.

IPGBs.StandardDecomposition.standard_decomposition — Method

Implementation of the standard decomposition algorithm introduced in Hosten and Thomas (1998).

IPGBs.StandardDecomposition.support_in — Method

Returns true iff the support of monomial is contained in the set represented by sigma.

IPGBs.StandardDecomposition.update_ideal! — Method

Updates the set of minimal generators of the ideal generated by gens upon adding new_gen.