SignatureAlgorithms.jl

Documentation for SignatureAlgorithms.jl

IPGBs.SignatureAlgorithms — Module

An implementation of a Signature-based algorithm for Gröbner bases of toric ideals.

IPGBs.SignatureAlgorithms.SignatureSet — Type

Stores one SupportTree for each signature index. This enables fast divisor queries for Signatures.

IPGBs.GBAlgorithms.initialize! — Method

Initializes all data in algorithm. Must be called before using any of the data inside for GB computations.

TODO almost all of this logic is still duplicated in Buchberger.jl. Can I do #this better?

IPGBs.GBAlgorithms.late_pair_elimination — Method

Attempts to eliminate pair before building it as an explicit GBElement, after retrieving it from the priority queue for processing.

Returns true iff pair may be eliminated at this point.

IPGBs.GBAlgorithms.next_pair! — Method

Generates the next pair to be processed by algorithm by picking the first element of its priority queue. Returns nothing if there are no more S-pairs to be processed and the algorithm may thus terminate.

Pairs are processed in order of increasing signature.

IPGBs.GBAlgorithms.process_zero_reduction! — Method

Adds syzygy_element to the set of known syzygies of algorithm.

IPGBs.GBAlgorithms.update! — Method

Adds the element g to the current basis maintained by algorithm. Uses pair, if available, to generate new Koszul syzygies. pair is assumed to be the SignaturePair that was used to generate g.

IPGBs.SignatureAlgorithms.base_divisors_criterion — Method

Applies the Base Divisors Criterion presented in Roune and Stillman (2012).

IPGBs.SignatureAlgorithms.divided_by — Method

Returns true iff there is a signature in signatures that divides s. Uses SupportTrees for efficient divisor search.

IPGBs.SignatureAlgorithms.early_pair_elimination — Method

Attempts to eliminate pair before putting it into the priority queue, but after computing its signature.

Returns true iff pair may be eliminated at this point.

IPGBs.SignatureAlgorithms.enumerate_divisors — Method

Returns a list of all divisors of s in signatures that are distinct from s.

IPGBs.SignatureAlgorithms.high_base_divisor — Method

Computes the index of the best high-ratio base divisor for the element of index j.

IPGBs.SignatureAlgorithms.high_base_divisors_criterion — Method

The Base divisor criterion for high-ratio elements. An element a is high-ratio if lt(a) | lt(b), where b is the new element being added.

IPGBs.SignatureAlgorithms.koszul_criterion — Method

Applies the Koszul criterion to spair, returning true iff it can be eliminated by this criterion. The main idea is checking whether there exists a regular Koszul syzygy with the same signature as spair. If this happens, spair may be eliminated.

TODO this isn't eliminating pretty much anything at all! Investigate

IPGBs.SignatureAlgorithms.low_base_divisors_criterion — Method

The Base divisor criterion for low-ratio elements. An element a is low-ratio if signature(a) | signature(b), where b is the new element being added.

IPGBs.SignatureAlgorithms.signature_criterion — Method

Returns true iff spair is eliminated by the signature criterion, i.e. there is a known syzygy that divides its signature.

IPGBs.SignatureAlgorithms.update_queue! — Method

Adds new SignaturePairs to algorithm.heap. Applies early elimination of S-pairs when possible.

IPGBs.SignatureAlgorithms.update_syzygies! — Method

Adds the Koszul syzygies corresponding to the newest element of gb to the syzygy list.

Currently unused, as this is inefficient. It's not necessary to add all Koszul syzygies.

IPGBs.SignatureAlgorithms.very_early_pair_elimination — Method

Attempts to eliminate the pair (i, j) before even creating it as a SignaturePair (thus, without computing its signature).

Returns true iff (i, j) may be eliminated at this point.